requirements.txt
langchain
langchain-community
langchain-core
langchain_groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
langchain-text-splitters